requirements.txt
python-telegram-bot==20.6
fastapi==0.111.0
uvicorn==0.30.1
Pillow==10.3.0
